requirements.txt
fastapi
pydantic
docker
pymongo
python-dotenv
uvicorn
python-multipart